+++ src/otx/algorithms/detection/adapters/mmdet/models/detectors/custom_deformable_detr_detector.py
@@ -90,7 +90,7 @@ if is_mmdeploy_enabled():
         if ctx.cfg["dump_features"]:
             feature_vector = FeatureVectorHook.func(feat)
             cls_scores = outs[0]
-            saliency_map = ActivationMapHook.func(cls_scores)
+            saliency_map = ActivationMapHook(self).func(cls_scores)
             return (*bbox_results, feature_vector, saliency_map)
 
         return bbox_results